third_party/WebKit/LayoutTests/external/wpt/fetch/api/request/request-reset-attributes.https-expected.txt

This is a testharness.js-based test.
FAIL Request.isReloadNavigation is reset with non-empty RequestInit assert_equals: expected "old: false, new: false" but got "hello\n"
Harness: the test ran to completion.
